requirements.txt
numpy
scipy
matplotlib
pandas
python-dateutil
pyinstaller
pytest
